ISPN-12623 Move CDI tests to integrationtests/server-integration/server-integration-commons project
