+++ integrationtests/server-integration/server-integration-commons/src/test/java/org/infinispan/test/integration/cdi/GreetingCacheManager.java
@@ -1,4 +1,4 @@
-package org.infinispan.test.integration.as.cdi;
+package org.infinispan.test.integration.cdi;
 
 
 import java.util.Collection;
@@ -18,7 +18,7 @@ import org.infinispan.eviction.EvictionType;
  * <p>This manager is used to collect informations on the greeting cache and to clear it's content if needed.</p>
  *
  * @author Kevin Pollet &lt;pollet.kevin@gmail.com&gt; (C) 2011
- * @see javax.cache.annotation.CacheRemoveAll
+ * @see CacheRemoveAll
  */
 @Named
 @ApplicationScoped
