+++ integrationtests/server-integration/server-integration-commons/src/test/resources/beans.xml
