+++ integrationtests/server-integration/wildfly-modules/src/test/java/org/infinispan/test/integration/as/cdi/GreetingCacheManagerIT.java
@@ -1,11 +1,7 @@
 package org.infinispan.test.integration.as.cdi;
 
-import static org.junit.Assert.assertEquals;
-
-import javax.inject.Inject;
-
 import org.infinispan.commons.util.Version;
-import org.infinispan.eviction.EvictionType;
+import org.infinispan.test.integration.cdi.AbstractGreetingCacheManagerIT;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.shrinkwrap.api.Archive;
@@ -15,21 +11,19 @@ import org.jboss.shrinkwrap.api.asset.StringAsset;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.jboss.shrinkwrap.descriptor.api.Descriptors;
 import org.jboss.shrinkwrap.descriptor.api.spec.se.manifest.ManifestDescriptor;
-import org.junit.Before;
-import org.junit.Test;
 import org.junit.runner.RunWith;
 
 /**
  * @author Kevin Pollet &lt;pollet.kevin@gmail.com&gt; (C) 2011
  */
 @RunWith(Arquillian.class)
-public class GreetingCacheManagerIT {
+public class GreetingCacheManagerIT extends AbstractGreetingCacheManagerIT {
 
    @Deployment
    public static Archive<?> deployment() {
       return ShrinkWrap
             .create(WebArchive.class, "cdi-cm.war")
-            .addPackage(GreetingCacheManagerIT.class.getPackage())
+            .addPackage(AbstractGreetingCacheManagerIT.class.getPackage())
             .add(manifest(), "META-INF/MANIFEST.MF")
             .addAsWebInfResource("beans.xml");
    }
@@ -38,49 +32,5 @@ public class GreetingCacheManagerIT {
       String manifest = Descriptors.create(ManifestDescriptor.class)
             .attribute("Dependencies", "org.infinispan:" + Version.getModuleSlot() + " services").exportAsString();
       return new StringAsset(manifest);
-   }
-
-   @Inject
-   private GreetingService greetingService;
-
-   @Inject
-   private GreetingCacheManager greetingCacheManager;
-
-   @Before
-   public void init() {
-      greetingCacheManager.clearCache();
-      assertEquals(0, greetingCacheManager.getNumberOfEntries());
-   }
-
-   @Test
-   public void testGreetingCacheConfiguration() {
-      // Cache name
-      assertEquals("greeting-cache", greetingCacheManager.getCacheName());
-
-      // Eviction
-      assertEquals(128, greetingCacheManager.getMemorySize());
-      assertEquals(EvictionType.COUNT, greetingCacheManager.getEvictionType());
-
-      // Lifespan
-      assertEquals(-1, greetingCacheManager.getExpirationLifespan());
-   }
-
-   @Test
-   public void testGreetingCacheCachedValues() {
-      greetingService.greet("Pete");
-
-      assertEquals(1, greetingCacheManager.getCachedValues().length);
-      assertEquals("Hello Pete :)", greetingCacheManager.getCachedValues()[0]);
-   }
-
-   @Test
-   public void testClearGreetingCache() {
-      greetingService.greet("Pete");
-
-      assertEquals(1, greetingCacheManager.getNumberOfEntries());
-
-      greetingCacheManager.clearCache();
-
-      assertEquals(0, greetingCacheManager.getNumberOfEntries());
    }
 }
